created wish interface and changed naming
