+++ src/main/java/com/group6/wishshare/data/repository/UserRepositoryStub.java
@@ -1,12 +1,11 @@
 package com.group6.wishshare.data.repository;
 
 import com.group6.wishshare.domain.model.User;
-import com.group6.wishshare.domain.service.DataFacadeInterface;
 import com.group6.wishshare.domain.service.LoginException;
 
-public class DataFacade implements DataFacadeInterface {
+public class UserRepositoryStub implements UserRepository {
 
-  private final UserRepository USER_REPOSITORY = new UserRepository();
+  private final UserRepositoryImpl USER_REPOSITORY = new UserRepositoryImpl();
 
   public User login(String email, String password) throws LoginException {
     return USER_REPOSITORY.login(email, password);
