+++ src/main/java/com/group6/wishshare/data/repository/WishRepositoryStub.java
@@ -1,0 +1,4 @@
+package com.group6.wishshare.data.repository;
+
+public class WishRepositoryStub {
+}
